Review and purchased car entity is added

+++ car-purcharse-service/src/main/java/com/hcltech/car_purcharse_service/repository/BuyerRepository.java
@@ -3,4 +3,4 @@ package com.hcltech.car_purcharse_service.repository;
 import com.hcltech.car_purcharse_service.model.Buyer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BuyerRepository extends JpaRepository<Buyer, Long> {}
+public interface BuyerRepository extends JpaRepository<Buyer, Integer> {}
